refactor some functionality and css sizes

+++ src/components/TodosInfo.js
@@ -14,7 +14,7 @@ import styles from '../styles/TodosInfo.module.css';
 import TodosSort from './TodosSort';
 import { toast } from 'react-toastify';
 
-const TodosInfo = ({ setTodos, sort, setSort }) => {
+const TodosInfo = ({ sort, setSort }) => {
 	const [remainingTodos, setRemainingTodos] = useState(0);
 
 	// redux
@@ -38,36 +38,55 @@ const TodosInfo = ({ setTodos, sort, setSort }) => {
 		});
 	}, []);
 
-	const clearCompletedTodos = () => {
+	const clearCompletedTodos = async () => {
 		const q = query(
 			colRef,
 			where('userId', '==', auth?.currentUser.uid),
 			where('isDone', '==', true)
 		);
 
-		getDocs(q)
-			.then((snapshot) => {
+		const querySnapshot = await getDocs(
+			query(colRef, where('isDone', '==', true))
+		);
+
+		const completedTodos = [];
+
+		querySnapshot.forEach((doc) => {
+			// doc.data() is never undefined for query doc snapshots
+			completedTodos.push({ todoId: doc.id, data: doc.data() });
+		});
+
+		if (completedTodos.length >= 1) {
+			getDocs(q).then((snapshot) => {
 				snapshot.docs.map((doc) => {
 					deleteDoc(doc.ref);
 				});
-			})
-			.then(() => toast('Completed Todos Cleared'));
+				toast.success('Completed Todos Cleared');
+			});
+		}
 	};
 
 	return (
 		// todo information
-		<div
-			className={`${styles.todosInfo} ${darkMode && styles['dark-todosInfo']}`}
-		>
-			<p className={styles.left}>{remainingTodos.length} Todos Remaining</p>
+		<>
+			<div
+				className={`${styles.todosInfo} ${
+					darkMode && styles['dark-todosInfo']
+				}`}
+			>
+				<p className={styles.left}>
+					{remainingTodos.length}{' '}
+					{remainingTodos.length <= 1 ? 'Todo Remaining' : 'Todos Remaining'}
+				</p>
 
-			{/* todos sort */}
-			<div className={styles.todosSort}>
-				<TodosSort setTodos={setTodos} sort={sort} setSort={setSort} />
+				{/* todos sort */}
+				<div className={styles.todosSort}>
+					<TodosSort sort={sort} setSort={setSort} />
+				</div>
+
+				<p onClick={clearCompletedTodos}>Clear Completed</p>
 			</div>
-
-			<p onClick={clearCompletedTodos}>Clear Completed</p>
-		</div>
+		</>
 	);
 };
